Rename algs -> samplers
