+++ salmon/triplets/samplers/_random_sampling.py
