+++ salmon/triplets/samplers/adaptive/_embed.py
