+++ salmon/triplets/samplers/adaptive/_noise_models.py
