+++ salmon/triplets/samplers/adaptive/search/__init__.py
