+++ salmon/triplets/samplers/adaptive/search/_search.py
