+++ salmon/triplets/samplers/adaptive/search/tests/utilsSTE.py
