+++ salmon/triplets/samplers/utils.py
